move coreos version handling into it's own package
